requirements.txt
asgiref==3.8.1
blinker==1.4
certifi==2024.2.2
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
filelock==3.14.0
gyp==0.1
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
pytz==2024.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.0
systemd-python==234
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.26.2
virtualenv-clone==0.3.0
wadllib==1.3.6
whitenoise==6.6.0
zipp==1.0.0
